Split tests up just like the files
