+++ core/test/unit/permissions/index_spec.js
@@ -1,12 +1,11 @@
 var should = require('should'),
     sinon = require('sinon'),
-    testUtils = require('../utils'),
+    testUtils = require('../../utils'),
     Promise = require('bluebird'),
     _ = require('lodash'),
-    Models = require('../../server/models'),
-    errors = require('../../server/errors'),
-    permissions = require('../../server/permissions'),
-    parseContext = require('../../server/permissions/parse-context'),
+    Models = require('../../../server/models'),
+    errors = require('../../../server/errors'),
+    permissions = require('../../../server/permissions'),
 
     sandbox = sinon.sandbox.create();
 
@@ -37,250 +36,6 @@ describe.only('Permissions', function () {
                 actionsMap.edit.sort().should.eql(['post', 'tag', 'user', 'page'].sort());
 
                 actionsMap.should.equal(permissions.actionsMap);
-
-                done();
-            }).catch(done);
-        });
-    });
-
-    describe('parseContext', function () {
-        it('should return public for no context', function () {
-            parseContext().should.eql({
-                internal: false,
-                external: false,
-                user: null,
-                app: null,
-                public: true
-            });
-            parseContext({}).should.eql({
-                internal: false,
-                external: false,
-                user: null,
-                app: null,
-                public: true
-            });
-        });
-
-        it('should return public for random context', function () {
-            parseContext('public').should.eql({
-                internal: false,
-                external: false,
-                user: null,
-                app: null,
-                public: true
-            });
-            parseContext({client: 'thing'}).should.eql({
-                internal: false,
-                external: false,
-                user: null,
-                app: null,
-                public: true
-            });
-        });
-
-        it('should return user if user populated', function () {
-            parseContext({user: 1}).should.eql({
-                internal: false,
-                external: false,
-                user: 1,
-                app: null,
-                public: false
-            });
-        });
-
-        it('should return app if app populated', function () {
-            parseContext({app: 5}).should.eql({
-                internal: false,
-                external: false,
-                user: null,
-                app: 5,
-                public: false
-            });
-        });
-
-        it('should return internal if internal provided', function () {
-            parseContext({internal: true}).should.eql({
-                internal: true,
-                external: false,
-                user: null,
-                app: null,
-                public: false
-            });
-
-            parseContext('internal').should.eql({
-                internal: true,
-                external: false,
-                user: null,
-                app: null,
-                public: false
-            });
-        });
-
-        it('should return external if external provided', function () {
-            parseContext({external: true}).should.eql({
-                internal: false,
-                external: true,
-                user: null,
-                app: null,
-                public: false
-            });
-
-            parseContext('external').should.eql({
-                internal: false,
-                external: true,
-                user: null,
-                app: null,
-                public: false
-            });
-        });
-    });
-
-    describe('applyPublicRules', function () {
-        it('should return empty object for docName with no rules', function (done) {
-            permissions.applyPublicRules('test', 'test', {}).then(function (result) {
-                result.should.eql({});
-                done();
-            });
-        });
-
-        it('should return unchanged object for non-public context', function (done) {
-            var internal = {context: 'internal'},
-                user = {context: {user: 1}},
-                app = {context: {app: 1}};
-
-            permissions.applyPublicRules('posts', 'browse', _.cloneDeep(internal)).then(function (result) {
-                result.should.eql(internal);
-
-                return permissions.applyPublicRules('posts', 'browse', _.cloneDeep(user));
-            }).then(function (result) {
-                result.should.eql(user);
-
-                return permissions.applyPublicRules('posts', 'browse', _.cloneDeep(app));
-            }).then(function (result) {
-                result.should.eql(app);
-
-                done();
-            }).catch(done);
-        });
-
-        it('should return unchanged object for post with public context', function (done) {
-            var publicContext = {context: {}};
-
-            permissions.applyPublicRules('posts', 'browse', _.cloneDeep(publicContext)).then(function (result) {
-                result.should.not.eql(publicContext);
-                result.should.eql({
-                    context: {},
-                    status: 'published'
-                });
-
-                return permissions.applyPublicRules('posts', 'browse', _.extend({}, _.cloneDeep(publicContext), {status: 'published'}));
-            }).then(function (result) {
-                result.should.eql({
-                    context: {},
-                    status: 'published'
-                });
-
-                done();
-            }).catch(done);
-        });
-
-        it('should throw an error for draft post without uuid (read)', function (done) {
-            var draft = {context: {}, data: {status: 'draft'}};
-
-            permissions.applyPublicRules('posts', 'read', _.cloneDeep(draft)).then(function () {
-                done('Did not throw an error for draft');
-            }).catch(function (err) {
-                (err instanceof errors.NoPermissionError).should.eql(true);
-                done();
-            });
-        });
-
-        it('should throw an error for draft post (browse)', function (done) {
-            var draft = {context: {}, status: 'draft'};
-
-            permissions.applyPublicRules('posts', 'browse', _.cloneDeep(draft)).then(function () {
-                done('Did not throw an error for draft');
-            }).catch(function (err) {
-                (err instanceof errors.NoPermissionError).should.eql(true);
-                done();
-            });
-        });
-
-        it('should permit post draft status with uuid (read)', function (done) {
-            var draft = {context: {}, data: {status: 'draft', uuid: '1234-abcd'}};
-
-            permissions.applyPublicRules('posts', 'read', _.cloneDeep(draft)).then(function (result) {
-                result.should.eql(draft);
-                done();
-            }).catch(done);
-        });
-
-        it('should permit post all status with uuid (read)', function (done) {
-            var draft = {context: {}, data: {status: 'all', uuid: '1234-abcd'}};
-
-            permissions.applyPublicRules('posts', 'read', _.cloneDeep(draft)).then(function (result) {
-                result.should.eql(draft);
-                done();
-            }).catch(done);
-        });
-
-        it('should NOT permit post draft status with uuid (browse)', function (done) {
-            var draft = {context: {}, status: 'draft', uuid: '1234-abcd'};
-
-            permissions.applyPublicRules('posts', 'browse', _.cloneDeep(draft)).then(function () {
-                done('Did not throw an error for draft');
-            }).catch(function (err) {
-                (err instanceof errors.NoPermissionError).should.eql(true);
-                done();
-            });
-        });
-
-        it('should NOT permit post all status with uuid (browse)', function (done) {
-            var draft = {context: {}, status: 'all', uuid: '1234-abcd'};
-
-            permissions.applyPublicRules('posts', 'browse', _.cloneDeep(draft)).then(function () {
-                done('Did not throw an error for draft');
-            }).catch(function (err) {
-                (err instanceof errors.NoPermissionError).should.eql(true);
-                done();
-            });
-        });
-
-        it('should throw an error for draft post with uuid and id or slug (read)', function (done) {
-            var draft = {context: {}, data: {status: 'draft', uuid: '1234-abcd', id: 1}};
-
-            permissions.applyPublicRules('posts', 'read', _.cloneDeep(draft)).then(function () {
-                done('Did not throw an error for draft');
-            }).catch(function (err) {
-                (err instanceof errors.NoPermissionError).should.eql(true);
-
-                draft = {context: {}, data: {status: 'draft', uuid: '1234-abcd', slug: 'abcd'}};
-
-                return permissions.applyPublicRules('posts', 'read', _.cloneDeep(draft)).then(function () {
-                    done('Did not throw an error for draft');
-                }).catch(function (err) {
-                    (err instanceof errors.NoPermissionError).should.eql(true);
-                    done();
-                });
-            });
-        });
-
-        it('should return unchanged object for user with public context', function (done) {
-            var publicContext = {context: {}};
-
-            permissions.applyPublicRules('users', 'browse', _.cloneDeep(publicContext)).then(function (result) {
-                result.should.not.eql(publicContext);
-                result.should.eql({
-                    context: {},
-                    status: 'all'
-                });
-
-                return permissions.applyPublicRules('users', 'browse', _.extend({}, _.cloneDeep(publicContext), {status: 'active'}));
-            }).then(function (result) {
-                result.should.eql({
-                    context: {},
-                    status: 'active'
-                });
 
                 done();
             }).catch(done);
